pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "langevals"
version = "0.1.1"
description = "A platform to evaluate LLM outputs using various evaluators."
authors = [ "Rogerio Chaves <rogerio@langwatch.ai>",]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
hypercorn = "^0.16.0"
pydantic = "^2.6.3"
python-dotenv = "1.0.1"

[tool.poetry.scripts]
langevals-server = "langevals.server:main"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
toml = "^0.10.2"
grevillea = "^0.0.1"

[tool.poetry.extras]
azure = [ "langevals-azure",]
openai = [ "langevals-openai",]
example = [ "langevals-example",]
ragas = [ "langevals-ragas",]
google_cloud = [ "langevals-google_cloud",]

[tool.poetry.dependencies.langevals-core]
path = "langevals_core"
develop = true

[tool.poetry.dependencies.langevals-azure]
path = "evaluators/azure"
develop = true
optional = true

[tool.poetry.dependencies.langevals-openai]
path = "evaluators/openai"
develop = true
optional = true

[tool.poetry.dependencies.langevals-example]
path = "evaluators/example"
develop = true
optional = true

[tool.poetry.dependencies.langevals-ragas]
path = "evaluators/ragas"
develop = true
optional = true

[tool.poetry.dependencies.langevals-google_cloud]
path = "evaluators/google_cloud"
develop = true
optional = true
